Use cheerio to parse/patch HTML in _convertImagesToInline; data-filename
Instead of using regexes to parse and edit HTML, use a DOM parser that
works a bit better (e.g. HTML entities are converted)

Also add support for an optional `data-filename` attribute so that
the image attachment filename can be renamed -- useful if the image
has a long or non-semantic filename.